move aiderBuffer related functions to aiderCommand.ts
